src/test/java/com/maroon5mlj/service/impl/ProductCategoryServiceImplTest.java
package com.maroon5mlj.service.impl;

import com.maroon5mlj.dataobject.ProductCategory;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.Date;
import java.util.List;

import static org.junit.Assert.*;

/**
 * Created by lovea on 2017/10/27.
 */
@RunWith(SpringRunner.class)
@SpringBootTest
public class ProductCategoryServiceImplTest {

    @Autowired
    private ProductCategoryServiceImpl productCategoryService;

    @Test
    public void findOne() throws Exception {
        ProductCategory productCategory = productCategoryService.findOne(1);
        Assert.assertEquals(new Integer(1),productCategory.getCategoryId());
    }

    @Test
    public void findAll() throws Exception {
        List<ProductCategory> resultList = productCategoryService.findAll();
        Assert.assertNotEquals(0,resultList.size());
    }

    @Test
    public void findByCategoryTypeIn() throws Exception {
        List<ProductCategory> resultList = productCategoryService.findByCategoryTypeIn(Arrays.asList(1,2,3));
        Assert.assertNotEquals(0,resultList.size());
    }

    @Test
    public void save() throws Exception {
        ProductCategory productCategory = productCategoryService.save(new ProductCategory("测试", 5,new Date()));

        Assert.assertNotEquals(null,productCategory);
    }

}